02_Applications/00_aurex-platform/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.8

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0

# Messaging
aio-pika==9.3.1
celery==5.3.4
kombu==5.3.4

# Utilities
httpx==0.25.2
orjson==3.9.10
email-validator==2.1.0
python-dateutil==2.8.2

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging
loguru==0.7.2
structlog==23.2.0